refactor: split DirectoryViewTable, consolidate column-filter state (C-104)
Split the DirectoryViewTable dispatcher into two specialized components
and consolidate column-filter state on the Zustand store:

- DirectoryViewTableConnection: bucket-specific columns + cell renderers
- DirectoryViewTableDirectory: file/directory-specific columns + cell renderers
- Delete DirectoryViewTable.tsx (dispatcher) — callers pick directly
- Split the dispatcher's tests into per-component test files
- Remove the controlled-mode escape hatch from useColumnFilters / Table:
  all consumers now read/write the per-tableId Zustand store directly.
  recent.route no longer keeps local useState for column filters.

TODO(C-82) markers added to DirectoryView and recent.route flagging the
remaining bridge where Grid/Tree views need filter state read externally.
True fix is pushing filter awareness into each view (C-82).
