+++ app/components/Table/types.ts
@@ -1,6 +1,5 @@
 import type {
   ColumnDefBase,
-  ColumnFiltersState,
   FilterFn,
   OnChangeFn,
   RowSelectionState,
@@ -78,7 +77,5 @@ export interface TableProps<TData extends object> {
   rowSelection?: RowSelectionState;
   onRowSelectionChange?: OnChangeFn<RowSelectionState>;
   getRowId?: (row: TData) => string;
-  columnFilters?: ColumnFiltersState;
-  onColumnFiltersChange?: OnChangeFn<ColumnFiltersState>;
   showFilters?: boolean;
 }
